# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */
import type { BookingStatus } from './bookingStatus';

export interface Booking {
  id: number;
  clientId: number;
  providerId: number;
  serviceId: number;
  status: BookingStatus;
  scheduledAt: Date;
  address: string;
  city: string;
  postalCode?: string | null;
  careNotes?: string | null;
  clientNotes?: string | null;
  cancellationReason?: string | null;
  createdAt?: Date;
  updatedAt?: Date;
}