# lib/api-zod/src/generated/types/bookingListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */
import type { Booking } from './booking';

export interface BookingListResponse {
  bookings: Booking[];
  total: number;
  limit: number;
  offset: number;
}