# lib/api-zod/src/generated/types/bookingStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  requested: 'requested',
  confirmed: 'confirmed',
  completed: 'completed',
  cancelled: 'cancelled',
  rescheduled: 'rescheduled',
  no_show: 'no_show',
} as const;