# lib/api-zod/src/generated/types/createBookingRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */

export interface CreateBookingRequest {
  /** provider_profiles.id */
  providerId: number;
  serviceId: number;
  scheduledAt: Date;
  /** @minLength 1 */
  address: string;
  /** @minLength 1 */
  city: string;
  postalCode?: string;
  careNotes?: string;
  clientNotes?: string;
}