# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */
import type { InvoiceStatus } from './invoiceStatus';

export interface Invoice {
  id: number;
  bookingId: number;
  clientId: number;
  providerId: number;
  /** Amount in cents (CAD) */
  amountCents: number;
  status: InvoiceStatus;
  paidAt?: Date | null;
  createdAt?: Date;
  updatedAt?: Date;
}