# lib/api-zod/src/generated/types/invoiceResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */
import type { Invoice } from './invoice';

export interface InvoiceResponse {
  invoice: Invoice;
}