# lib/api-zod/src/generated/types/invoiceStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */

export type InvoiceStatus = typeof InvoiceStatus[keyof typeof InvoiceStatus];


export const InvoiceStatus = {
  pending: 'pending',
  paid: 'paid',
  cancelled: 'cancelled',
} as const;