# lib/api-zod/src/generated/types/updateBookingStatusRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * OnCall Foot API — foot rejuvenation marketplace
 * OpenAPI spec version: 0.3.0
 */
import type { BookingStatus } from './bookingStatus';

export interface UpdateBookingStatusRequest {
  status: BookingStatus;
  /** Required when transitioning to cancelled */
  cancellationReason?: string;
  /** Required when transitioning to rescheduled */
  scheduledAt?: Date;
}